move SynchronizedMutableMap to synchronous package
