+++ library/manager/kmp-tor-manager/src/commonMain/kotlin/io/matthewnelson/kmp/tor/manager/internal/util/synchronous/SynchronizedMutableMap.kt
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.manager.internal.util
+package io.matthewnelson.kmp.tor.manager.internal.util.synchronous
 
 import kotlinx.atomicfu.locks.SynchronizedObject
 import kotlinx.atomicfu.locks.synchronized
 
-internal class SynchronizedMutableMap<T: Any>: SynchronizedObject() {
+internal class SynchronizedMutableMap<T: Any>(initialCapacity: Int = 1): SynchronizedObject() {
 
-    private val map: MutableMap<String, T> = LinkedHashMap(1)
+    private val map: MutableMap<String, T> = LinkedHashMap(initialCapacity)
 
-    fun <V: Any?> withLock(block: MutableMap<String, T>.() -> V): V {
+    @JvmSynthetic
+    internal fun <V: Any?> withLock(block: MutableMap<String, T>.() -> V): V {
         return synchronized(this) {
             block.invoke(map)
         }
